+++ controllers/sensor/trigger.go
@@ -33,7 +33,7 @@ func (sc *sensorExecutionCtx) executeTrigger(trigger v1alpha1.Trigger) error {
 		if err != nil {
 			return err
 		}
-		reader, err := store.GetArtifactReader(&trigger.Resource.Source, creds)
+		reader, err := store.GetArtifactReader(&trigger.Resource.Source, creds, sc.kubeClient)
 		if err != nil {
 			return err
 		}